requirements.txt
# FastAPI and WebSocket
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
websockets==12.0

# ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python==4.8.1.78
Pillow>=10.0.0
numpy>=1.24.0

# Data Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async and Utilities
asyncio-mqtt==0.13.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog>=23.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Production
gunicorn>=21.2.0 